Automatically backed up by Learn

+++ lib/student.rb
@@ -16,7 +16,7 @@ class Student
       CREATE TABLE IF NOT EXISTS songs (
         id INTEGER PRIMARY KEY, 
         name TEXT, 
-        grade INTEGER
+        grade NUMERIC
         )
         SQL
     DB[:conn].execute(sql)
